@@ -6,7 +6,7 @@ authors = ["timoniq <tesseradecades@mail.ru>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.11"
 aiohttp = "^3.8.1"
 vbml = "^1.1.post1"
 choicelib = "^0.1.5"
@@ -15,6 +15,7 @@ PyYAML = "^6.0"
 certifi = "^2022.6.15"
 msgspec = "^0.9.0"
 requests = "^2.28.1"
+# typing-extensions = "^4.5.0"
 
 [tool.poetry.dev-dependencies]
 
