@@ -21,14 +21,14 @@ class MessageCute(Message):
 
     async def answer(
         self,
-        text: typing.Optional[str] = None,
-        parse_mode: typing.Optional[str] = None,
-        entities: typing.Optional[typing.List["MessageEntity"]] = None,
-        disable_web_page_preview: typing.Optional[bool] = None,
-        disable_notification: typing.Optional[bool] = None,
-        protect_content: typing.Optional[bool] = None,
-        reply_to_message_id: typing.Optional[int] = None,
-        allow_sending_without_reply: typing.Optional[bool] = None,
+        text: str | None = None,
+        parse_mode: str | None = None,
+        entities: list["MessageEntity"] | None = None,
+        disable_web_page_preview: bool | None = None,
+        disable_notification: bool | None = None,
+        protect_content: bool | None = None,
+        reply_to_message_id: int | None = None,
+        allow_sending_without_reply: bool | None = None,
         reply_markup: typing.Optional[
             typing.Union[
                 "InlineKeyboardMarkup",
@@ -46,13 +46,13 @@ class MessageCute(Message):
 
     async def reply(
         self,
-        text: typing.Optional[str] = None,
-        parse_mode: typing.Optional[str] = None,
-        entities: typing.Optional[typing.List["MessageEntity"]] = None,
-        disable_web_page_preview: typing.Optional[bool] = None,
-        disable_notification: typing.Optional[bool] = None,
-        protect_content: typing.Optional[bool] = None,
-        allow_sending_without_reply: typing.Optional[bool] = None,
+        text: str | None = None,
+        parse_mode: str | None = None,
+        entities: list["MessageEntity"] | None = None,
+        disable_web_page_preview: bool | None = None,
+        disable_notification: bool | None = None,
+        protect_content: bool | None = None,
+        allow_sending_without_reply: bool | None = None,
         reply_markup: typing.Optional[
             typing.Union[
                 "InlineKeyboardMarkup",
@@ -67,9 +67,7 @@ class MessageCute(Message):
         if "message_thread_id" not in params and self.is_topic_message:
             params["message_thread_id"] = self.message_thread_id
         return await self.ctx_api.send_message(
-            chat_id=self.chat.id,
-            reply_to_message_id=self.message_id,
-            **params
+            chat_id=self.chat.id, reply_to_message_id=self.message_id, **params
         )
 
     async def delete(self, **other) -> Result[bool, APIError]:
@@ -77,25 +75,21 @@ class MessageCute(Message):
         if "message_thread_id" not in params and self.is_topic_message:
             params["message_thread_id"] = self.message_thread_id
         return await self.ctx_api.delete_message(
-            chat_id=self.chat.id,
-            message_id=self.message_id,
-            **params
+            chat_id=self.chat.id, message_id=self.message_id, **params
         )
 
     async def edit(
         self,
-        text: typing.Optional[str] = None,
-        parse_mode: typing.Optional[str] = None,
-        entities: typing.Optional[typing.List[MessageEntity]] = None,
-        disable_web_page_preview: typing.Optional[bool] = None,
-        reply_markup: typing.Optional[InlineKeyboardMarkup] = None,
+        text: str | None = None,
+        parse_mode: str | None = None,
+        entities: list[MessageEntity] | None = None,
+        disable_web_page_preview: bool | None = None,
+        reply_markup: InlineKeyboardMarkup | None = None,
         **other
     ) -> Result[Message | bool, APIError]:
         params = get_params(locals())
         if "message_thread_id" not in params and self.is_topic_message:
             params["message_thread_id"] = self.message_thread_id
         return await self.ctx_api.edit_message_text(
-            chat_id=self.chat.id,
-            message_id=self.message_id,
-            **params
+            chat_id=self.chat.id, message_id=self.message_id, **params
         )
