@@ -2,6 +2,7 @@ import typing
 import types
 from telegrinder.bot.rules import ABCRule
 from telegrinder.tools import magic_bundle
+from telegrinder.bot.dispatch.process import check_rule
 from telegrinder.types import Update
 from telegrinder.api.abc import ABCAPI
 from .abc import ABCHandler
@@ -13,10 +14,10 @@ T = typing.TypeVar("T")
 class FuncHandler(ABCHandler, typing.Generic[T]):
     def __init__(
         self,
-        func: typing.Union[types.FunctionType, typing.Callable],
-        rules: typing.List[ABCRule],
+        func: types.FunctionType | typing.Callable,
+        rules: list[ABCRule],
         is_blocking: bool = True,
-        dataclass: typing.Optional[typing.Any] = dict,
+        dataclass: typing.Any | None = dict,
     ):
         self.func = func
         self.is_blocking = is_blocking
@@ -27,18 +28,7 @@ class FuncHandler(ABCHandler, typing.Generic[T]):
     async def check(self, api: ABCAPI, event: Update) -> bool:
         self.ctx = {}
         for rule in self.rules:
-            e = event
-            if rule.__event__ is None:
-                pass
-            else:
-                event_dict = event.to_dict()
-                if event_dict.get(rule.__event__.name) is None:
-                    continue
-                e = rule.__event__.dataclass(
-                    **event_dict[rule.__event__.name].to_dict(),
-                    api=api,
-                )
-            if not await rule.run_check(e, self.ctx):
+            if not await check_rule(api, rule, event, self.ctx):
                 logger.debug("Rule {} failed", rule)
                 return False
         return True
