@@ -8,5 +8,5 @@ class ABCMiddleware(ABC, typing.Generic[T]):
     async def pre(self, event: T, ctx: dict) -> bool:
         ...
 
-    async def post(self, event: T, responses: typing.List[typing.Any], ctx: dict):
+    async def post(self, event: T, responses: list[typing.Any], ctx: dict):
         ...
