@@ -5,16 +5,22 @@ from .middleware.abc import ABCMiddleware
 from .handler.abc import ABCHandler
 from telegrinder.types import Update
 from telegrinder.modules import logger
+from telegrinder.api.abc import ABCAPI
+
+if typing.TYPE_CHECKING:
+    from telegrinder.bot.rules.abc import ABCRule
 
 T = typing.TypeVar("T")
 E = typing.TypeVar("E")
+_ = typing.Any
 
 
 async def process_waiters(
-    waiters: typing.Dict[T, Waiter],
+    api: ABCAPI,
+    waiters: dict[T, Waiter],
     key: T,
-    event: typing.Optional[E],
-    raw_event: dict,
+    event: E | None,
+    raw_event: Update,
     str_handler: typing.Callable,
 ) -> bool:
     if key not in waiters:
@@ -28,10 +34,7 @@ async def process_waiters(
     ctx = {}
 
     for rule in waiter.rules:
-        chk_event = event
-        if rule.__event__ is None:
-            chk_event = raw_event
-        if not await rule.run_check(chk_event, ctx):
+        if not await check_rule(api, rule, raw_event, ctx):
             if not waiter.default:
                 return True
             elif isinstance(waiter.default, str):
@@ -51,8 +54,8 @@ async def process_waiters(
 async def process_inner(
     event: T,
     raw_event: Update,
-    middlewares: typing.List[ABCMiddleware[T]],
-    handlers: typing.List[ABCHandler[T]],
+    middlewares: list[ABCMiddleware[T]],
+    handlers: list[ABCHandler[T]],
 ) -> bool:
     logger.debug("processing {}", event.__class__.__name__)
     ctx = {}
@@ -77,3 +80,23 @@ async def process_inner(
         await middleware.post(event, responses, ctx)
 
     return found
+
+
+async def check_rule(
+    api: ABCAPI, rule: "ABCRule", update: Update, ctx: dict[str, _]
+) -> bool:
+    """Checks requirements, adapts update
+    Returns check result"""
+
+    ctx_copy = ctx.copy()
+
+    model = await rule.adapter.adapt(api, update)
+    if not model.is_ok:
+        return False
+
+    for requirement in rule.require:
+        if not await check_rule(api, requirement, update, ctx_copy):
+            return False
+
+    ctx.update(ctx_copy)
+    return await rule.check(model.unwrap(), ctx)
