@@ -14,14 +14,16 @@ import typing
 class CallbackQueryView(ABCView, WithWaiter[int, CallbackQueryCute]):
     def __init__(self):
         self.auto_rules: list[ABCRule] = []
-        self.handlers: typing.List[ABCHandler[CallbackQueryCute]] = []
-        self.middlewares: typing.List[ABCMiddleware[CallbackQueryCute]] = []
-        self.short_waiters: typing.Dict[int, Waiter] = {}
+        self.handlers: list[ABCHandler[CallbackQueryCute]] = []
+        self.middlewares: list[ABCMiddleware[CallbackQueryCute]] = []
+        self.short_waiters: dict[int, Waiter] = {}
 
     def __call__(self, *rules: ABCRule, is_blocking: bool = True):
         def wrapper(func: typing.Callable[..., typing.Coroutine]):
             self.handlers.append(
-                FuncHandler(func, [*self.auto_rules, *rules], is_blocking, dataclass=None)
+                FuncHandler(
+                    func, [*self.auto_rules, *rules], is_blocking, dataclass=None
+                )
             )
             return func
 
@@ -34,7 +36,12 @@ class CallbackQueryView(ABCView, WithWaiter[int, CallbackQueryCute]):
         query = CallbackQueryCute(**event.callback_query.to_dict(), api=api)
 
         if await process_waiters(
-            self.short_waiters, query.message.message_id, query, event, query.answer
+            api,
+            self.short_waiters,
+            query.message.message_id,
+            query,
+            event,
+            query.answer,
         ):
             return
 
