@@ -1,7 +1,9 @@
 import asyncio
 import dataclasses
 import typing
-from telegrinder.bot.rules import ABCRule
+
+if typing.TYPE_CHECKING:
+    from telegrinder.bot.rules.abc import ABCRule
 
 
 DefaultWaiterHandler = typing.Callable[[typing.Any], typing.Coroutine]
@@ -11,27 +13,27 @@ E = typing.TypeVar("E")
 
 @dataclasses.dataclass
 class Waiter:
-    rules: typing.Iterable[ABCRule]
+    rules: typing.Iterable["ABCRule"]
     event: asyncio.Event
-    default: typing.Optional[typing.Union[DefaultWaiterHandler, str]] = None
+    default: DefaultWaiterHandler | str | None = None
 
 
-async def wait(waiter: Waiter) -> typing.Tuple[typing.Any, dict]:
+async def wait(waiter: Waiter) -> tuple[typing.Any, dict]:
     await waiter.event.wait()
     event, ctx = getattr(waiter.event, "e")
     return event, ctx
 
 
 class WithWaiter(typing.Generic[T, E]):
-    short_waiters: typing.Dict[T, Waiter]
-    auto_rules: list[ABCRule]
+    short_waiters: dict[T, Waiter]
+    auto_rules: list["ABCRule"]
 
     async def wait_for_answer(
         self,
         key: T,
-        *rules: ABCRule,
-        default: typing.Optional[typing.Union[DefaultWaiterHandler, str]] = None
-    ) -> typing.Tuple[E, dict]:
+        *rules: "ABCRule",
+        default: DefaultWaiterHandler | str | None = None
+    ) -> tuple[E, dict]:
         event = asyncio.Event()
         waiter = Waiter([*self.auto_rules, *rules], event, default)
         self.short_waiters[key] = waiter
