@@ -31,15 +31,15 @@ ALLOWED_UPDATES = [
 class Polling(ABCPolling):
     def __init__(
         self,
-        api: typing.Optional[ABCAPI] = None,
-        offset: typing.Optional[int] = None,
+        api: ABCAPI | None = None,
+        offset: int | None = None,
     ):
         self.api = api
         self.offset = offset or 0
         self._stop = False
         self.allowed_updates = ALLOWED_UPDATES
 
-    async def get_updates(self) -> typing.Optional[Raw]:
+    async def get_updates(self) -> Raw | None:
         raw_updates = await self.api.request_raw(
             "getUpdates",
             {"offset": self.offset, "allowed_updates": self.allowed_updates},
@@ -49,13 +49,13 @@ class Polling(ABCPolling):
             exit(6)
         return raw_updates.unwrap()
 
-    async def listen(self) -> typing.AsyncIterator[typing.List[Update]]:
+    async def listen(self) -> typing.AsyncIterator[list[Update]]:
         logger.debug("listening polling")
         while not self._stop:
             try:
                 updates = await self.get_updates()
-                updates_list: typing.List[Update] = msgspec.json.decode(
-                    updates, type=typing.List[Update]
+                updates_list: list[Update] = msgspec.json.decode(
+                    updates, type=list[Update]
                 )
                 if updates_list:
                     yield updates_list
