@@ -1,14 +1,12 @@
 from .abc import Message, patcher
-from .text import ABCTextMessageRule
+from .text import TextMessageRule
 import typing
 import vbml
 
-PatternLike = typing.Union[str, vbml.Pattern]
+PatternLike = str | vbml.Pattern
 
 
-def check_string(
-    patterns: typing.Union[PatternLike, typing.List[PatternLike]], s: str, ctx: dict
-) -> bool:
+def check_string(patterns: PatternLike | list[PatternLike], s: str, ctx: dict) -> bool:
     if s is None:
         return False
     for pattern in patterns:
@@ -21,8 +19,8 @@ def check_string(
     return False
 
 
-class Markup(ABCTextMessageRule):
-    def __init__(self, patterns: typing.Union[PatternLike, typing.List[PatternLike]]):
+class Markup(TextMessageRule):
+    def __init__(self, patterns: PatternLike | list[PatternLike]):
         if not isinstance(patterns, list):
             patterns = [patterns]
         self.patterns = [
