@@ -1,12 +1,12 @@
-from .text import ABCTextMessageRule, Message
+from .text import TextMessageRule, Message
 
 
-class Integer(ABCTextMessageRule):
+class Integer(TextMessageRule):
     async def check(self, message: Message, ctx: dict) -> bool:
         return message.text.isdigit()
 
 
-class IntegerInRange(ABCTextMessageRule, require=[Integer()]):
+class IntegerInRange(TextMessageRule, require=[Integer()]):
     def __init__(self, rng: range):
         self.rng = rng
 
